// frontend/src/features/gateways/models/gateway.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gateways API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Device } from "./device";
/**
 *
 * @export
 * @interface Gateway
 */
export interface Gateway {
  /**
   *
   * @type {string}
   * @memberof Gateway
   */
  _id: string;
  /**
   *
   * @type {string}
   * @memberof Gateway
   */
  serialNumber: string;
  /**
   *
   * @type {string}
   * @memberof Gateway
   */
  name: string;
  /**
   *
   * @type {string}
   * @memberof Gateway
   */
  ipAddress: string;
  /**
   *
   * @type {Array<Device>}
   * @memberof Gateway
   */
  devices: Array<Device>;
}